feat: add summary of determinant

+++ compile_tex.sh
@@ -1,19 +1,25 @@
 #!/bin/bash
-# 并行编译所有 .tex 文件，输出 PDF 到同级目录，并删除临时文件
+# 只编译有变更的 .tex 文件，并行输出 PDF 到同级目录，并删除临时文件
 
-find . -type f -name "*.tex" | while read texfile; do
-    texdir=$(dirname "$texfile")
-    echo "Compiling $texfile ..."
+# 获取已修改或新增的 .tex 文件
+changed_files=$(git ls-files -m -o --exclude-standard "*.tex")
 
-    # 使用后台任务 (&) 并行执行
-    xelatex -interaction=nonstopmode -output-directory="$texdir" "$texfile" &
+if [ -z "$changed_files" ]; then
+    echo " ⛔ No changed .tex files to compile. ⛔ "
+else
+    for texfile in $changed_files; do
+        texdir=$(dirname "$texfile")
+        echo "Compiling $texfile ..."
 
-done
+        # 并行编译
+        xelatex -interaction=nonstopmode -output-directory="$texdir" "$texfile" &
+    done
 
-# 等待所有后台任务完成
-wait
+    # 等待所有后台任务完成
+    wait
 
-# 删除所有临时文件
-find . -type f \( -name "*.aux" -o -name "*.log" -o -name "*.out" -o -name "*.toc" \) -delete
+    # 删除所有临时文件
+    find . -type f \( -name "*.aux" -o -name "*.log" -o -name "*.out" -o -name "*.toc" \) -delete
 
-echo "🎉🎉All .tex files compiled🎉🎉"
+    echo "🎉🎉All changed .tex files compiled🎉🎉"
+fi
